Added collecting objects.

+++ sprint3/problems/scores/solution/src/model/model_invariants.h
@@ -7,8 +7,8 @@ const double OFFSET = 0.4;
 const double EPSILON = 0.001;   // Точность рассчетов.
 const double INITIAL_DOG_VELOCITY = 1.0;
 const size_t INITIAL_BAG_CAPACITY = 3;
-const size_t DOG_WIDTH = 0.3;
-const size_t LOOT_WIDTH = 0.0;
-const size_t BASE_WIDTH = 0.5;
+const double DOG_WIDTH = 0.3;
+const double LOOT_WIDTH = 0.0;
+const double BASE_WIDTH = 0.5;
 
 }
